home page: events and deadlines
