+++ fec/fec/static/js/home.js
@@ -1,11 +1,9 @@
 'use strict';
 
-var upcomingEvents = require('./upcoming-events');
+var homepageEvents = require('./home-events');
 
 // accessible tabs for alt sidebar
 require('./vendor/tablist').init();
 
-/* Homepage Upcoming Events */
-
-// - What's Happening section
-new upcomingEvents.UpcomingEvents();
+// Home Page: Events and deadlines
+new homepageEvents.HomepageEvents();
